// src/app/inter-service/models/interservicemark.ts
export interface InterServiceMark {
    interServiceMarkId: number;
    courseDurationId:number;
    courseNameId:number;
    organizationNameId: string;
    documentId: number;
    courseTypeId: number;
    traineeNominationId: number;
    traineeId: number;
    coursePosition: number;
    obtaintMark: string;
    remarks:string;
    //menuPosition: string;
    isActive: boolean;
   
}